+++ @angular-generic-table/core/components/gt-pagination.component.ts
@@ -1,9 +1,9 @@
-import { Component, OnInit, Input } from '@angular/core';
+import {Component, OnInit, Input } from '@angular/core';
 import {GenericTableComponent} from './generic-table.component';
 
 @Component({
   selector: 'gt-pagination',
-  template: `<nav aria-label="Table navigation" *ngIf="genericTable.gtInfo">
+  template: `<nav aria-label="Table navigation" *ngIf="genericTable && genericTable.gtInfo && ready">
   <ul class="pagination" [ngClass]="gtClasses">
     <li class="page-item" [ngClass]="{'disabled' : genericTable.gtInfo.pageCurrent === 1 || genericTable.loading }"><a class="page-link" href="javascript:void(0);" (click)="genericTable.gtInfo.pageCurrent > 1 && genericTable.previousPage()" [attr.aria-label]="genericTable.gtTexts.paginatePrevious"><span aria-hidden="true">&laquo;</span><span class="sr-only">{{genericTable.gtTexts.paginatePrevious}}</span></a></li>
     <li class="page-item" [ngClass]="{'disabled' : genericTable.loading && genericTable.gtInfo.pageCurrent !== page, 'active' : genericTable.gtInfo.pageCurrent === page }" *ngFor="let page of genericTable.gtInfo.pageTotal | gtPaginationPipe:genericTable.gtInfo.pageCurrent"><a class="page-link" [tabindex]="page === true ? -1:0" href="javascript:void(0);" (click)="page === true ? '':genericTable.goToPage(page)">{{page === true ? '&hellip;':page}}</a></li>
@@ -17,14 +17,19 @@ export class GtPaginationComponent implements OnInit {
 
   @Input() genericTable:GenericTableComponent<any,any>;
   @Input() gtClasses: string;
-
-  constructor() { }
+  public ready:boolean = false;
 
   ngOnInit() {
-
+    this.genericTable.gtEvent.subscribe((res)=>{
+      if(res.name === 'gt-info' && res.value.pageTotal > 0){
+        this.ready = true;
+      }
+    })
   }
+
 }
 import { Pipe, PipeTransform } from '@angular/core';
+
 
 
 @Pipe({